# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Zenstats — 自托管网站分析服务，提供隐私友好的 Web 分析 API。\n\n## 认证方式\n- **BearerAuth**: 在 Authorization 头中传入 `Bearer \u003cJWT Token\u003e`，用于管理类 API\n- **APIKeyAuth**: 在 Authorization 头中传入 API Key，用于统计查询 API\n\n## 统一响应格式\n成功: `{ \"code\": 200, \"message\": \"success\", \"data\": ... }`\n失败: `{ \"code\": \u003cstatus\u003e, \"message\": \"...\", \"error\": \"...\" }`\n\n## API 分组\n- **认证** — 登录、注册、令牌刷新、密码重置、邮箱验证、系统初始化\n- **站点管理** — 站点 CRUD、域名验证、屏蔽规则（IP/域名/国家）\n- **统计分析** — 聚合指标、时间序列、维度细分、实时访客、CSV 导出\n- **漏斗分析** — 漏斗 CRUD + 转化率分析\n- **目标管理** — 转化目标 CRUD\n- **事件采集** — 前端埋点 SDK 上报端点\n- **数据导入** — GA4 历史数据导入与查询\n- **用户** — 自定义搜索引擎、子账号管理、配额查询、个人资料\n- **API Key** — API Key 创建、列表、删除\n- **管理员** — 用户/站点/套餐管理、系统配置、系统统计\n- **健康检查** — 服务及依赖连通性检查",
        "title": "Zenstats API",
        "termsOfService": "https://github.com/zenstats/zenstats",
        "contact": {
            "name": "Zenstats",
            "url": "https://github.com/zenstats/zenstats",
            "email": "wrpota@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/sites/{domain}/verification-status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定域名的验证状态信息，包括验证令牌（仅未验证时返回）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "站点验证"
                ],
                "summary": "获取站点验证状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "站点域名",
                        "name": "domain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.SiteVerificationStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sites/{domain}/verify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "通过获取域名下的验证文件来验证所有权",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "站点验证"
                ],
                "summary": "验证站点域名所有权",
                "parameters": [
                    {
                        "type": "string",
                        "description": "站点域名",
                        "name": "domain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或验证失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/profile": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新当前用户的显示名称。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新用户资料",
                "parameters": [
                    {
                        "description": "更新资料参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/quota": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户的套餐额度及使用情况。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户额度信息",
                "responses": {
                    "200": {
                        "description": "额度信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.UserQuotaInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrorResponse": {
            "description": "接口返回的错误响应信息",
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "error": {
                    "description": "错误详情",
                    "type": "string"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
        "response.SuccessResponse": {
            "description": "接口返回的成功响应信息",
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应信息",
                    "type": "string"
                }
            }
        },
...
        "types.SiteVerificationStatus": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "verification_token": {
                    "type": "string"
                },
                "verified_at": {
                    "type": "string"
                }
            }
        },
...
        "types.UpdateProfileRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
...
        "types.UserQuotaInfo": {
            "type": "object",
            "properties": {
                "current_api_keys": {
                    "type": "integer"
                },
                "current_monthly_events": {
                    "type": "integer"
                },
                "current_sites": {
                    "type": "integer"
                },
                "current_sub_accounts": {
                    "type": "integer"
                },
                "custom_search_engines": {
                    "type": "boolean"
                },
                "group_name": {
                    "type": "string"
                },
                "max_api_keys": {
                    "type": "integer"
                },
                "max_monthly_events": {
                    "type": "integer"
                },
                "max_sites": {
                    "type": "integer"
                },
                "max_sub_accounts": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "在 Authorization 头中传入 API Key（无需 Bearer 前缀）。在站点管理页面创建 API Key。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "在 Authorization 头中传入 `Bearer \u003cJWT Token\u003e`。获取方式：调用 /auth/login 或 /auth/register。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}